+++ src/js/btnUp.js
@@ -1,0 +1,29 @@
+import refs from "./refs";
+import throttle from 'lodash.throttle';
+
+refs.btnUpEl.addEventListener('click', scrollUp);
+
+function show() {
+  refs.btnUpEl.classList.remove('btn-up_hide');
+}
+
+function hide() {
+  refs.btnUpEl.classList.add('btn-up_hide');
+}
+
+window.addEventListener(
+  'scroll',
+  throttle(() => {
+    const scrollY = window.scrollY || document.documentElement.scrollTop;
+
+    scrollY > 400 ? show() : hide();
+  }, 500)
+);
+
+function scrollUp() {
+  window.scrollTo({
+    top: 0,
+    left: 0,
+    behavior: 'smooth',
+  });
+}
