+++ src/sass/components/_pagination.scss
@@ -22,7 +22,7 @@
       width: 25px;
       height: 25px;
       display: inline-flex;
-      background-color: gold;
+      background-color: var(--pagination-number-bg-color);
       color: white;
       border-radius: 5px;
       flex-direction: column;
@@ -30,7 +30,7 @@
       text-align: center;
       vertical-align: middle;
       &:hover {
-        background-color: gold;
+        background-color: var(--pagination-number-bg-color);
       }
       @media (min-width: 768px) {
         width: 40px;
