+++ src/sass/utils/_variables.scss
@@ -43,17 +43,20 @@
   --box-shadow-card: 0px 5px 10px 0px rgba(250, 209, 5, 0.5); // Просто добавив "Щоб було"
   
   --main-font-family: Roboto, sans-serif;
-  --animation: 250ms cubic-bezier(0.4, 0, 0.2, 1); // З минулого проєкту
+  --animation: 3000ms cubic-bezier(0.4, 0, 0.2, 1); // З минулого проєкту
   
   --body-bg-color: #ffffff;
   
-  //modal-tem
+  //modal-team
   --bg-modal-team: #efecec;
   --color-focus-team: #aff1ba;
   --team-shedov: #7a7979;
   --team-border: rgba(138, 136, 136, 0.6);
   --position-team: #ff6b08;
   --backdrop-team: rgba(0, 0, 0, 0.2);
+
+  --spinner-border-top-color: #fad105;
+  --spinner-2-border-top-color: #fad105;
 
   //footer
